+++ test/check/checkbox_test.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_fast_forms/flutter_fast_forms.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -19,7 +21,7 @@ void main() {
     expect(find.byType(CheckboxListTile), findsOneWidget);
     expect(find.text(widget.helperText!), findsOneWidget);
     expect(find.text(widget.labelText!), findsOneWidget);
-    expect(find.text(widget.titleText), findsOneWidget);
+    expect(find.text(widget.titleText!), findsOneWidget);
   });
 
   testWidgets('updates FastCheckbox', (WidgetTester tester) async {
@@ -62,5 +64,18 @@ void main() {
     await tester.pumpAndSettle();
 
     expect(errorTextFinder, findsOneWidget);
+  });
+
+  testWidgets('adapts FastCheckbox to iOS', (WidgetTester tester) async {
+    debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+
+    await tester.pumpWidget(buildCupertinoTestApp(
+      const FastCheckbox(name: 'switch', adaptive: true, titleText: 'title'),
+    ));
+
+    expect(find.byType(CupertinoFormRow), findsOneWidget);
+    expect(find.byType(CupertinoCheckbox), findsOneWidget);
+
+    debugDefaultTargetPlatformOverride = null;
   });
 }