+++ CHANGELOG.md
@@ -1,3 +1,78 @@
+## [11.0.0] - 05/16/2023
+
+* upgrades to Flutter `3.10.0`
+* upgrades to Dart `3.0.0`
+* upgrades sample to Material 3
+* adds adaptiveness to `FastCheckbox` via `CupertinoCheckbox`
+* adds all new styling properties to `FastCheckbox`, `FastRangeSlider`, `FastSlider` and `FastSwitch`
+* adds `textStyle` property to all date time fields
+* adds `inputDecorationTheme` property to `FastForm`
+* fixes styling issues when `enabled: true`  
+* fixes incorrect default content padding on iOS
+
+**Breaking Changes**
+
+* reduces the default `InputDecoration` of `FastFormField<T>`
+
+**old**
+```dart
+InputDecoration(
+  contentPadding: widget.contentPadding,
+  errorText: field.errorText,
+  helperText: widget.helperText,
+  labelText: widget.labelText,
+  labelStyle: TextStyle(
+    color: field.enabled
+      ? theme.textTheme.bodyLarge!.color
+      : theme.disabledColor,
+  ),
+  enabled: field.enabled,
+  disabledBorder: OutlineInputBorder(
+    borderSide: BorderSide(color: theme.disabledColor, width: 1),
+  ),
+  enabledBorder: OutlineInputBorder(
+    borderSide: BorderSide(color: Colors.grey[700]!, width: 1),
+  ),
+  focusedBorder: OutlineInputBorder(
+    borderSide: BorderSide(color: theme.primaryColor, width: 2),
+  ),
+  errorBorder: const OutlineInputBorder(
+    borderSide: BorderSide(color: Colors.red, width: 2),
+  ),
+  focusedErrorBorder: OutlineInputBorder(
+    borderSide: BorderSide(color: Colors.red[500]!, width: 2),
+  ),
+  filled: false,
+  fillColor: Colors.white,
+);
+```
+
+**new**
+```dart
+InputDecoration(
+  contentPadding: widget.contentPadding ?? const EdgeInsets.fromLTRB(12.0, 8.0, 12.0, 8.0),
+  enabled: enabled,
+  errorText: errorText, 
+  helperText: widget.helperText,
+  labelText: widget.labelText,
+);
+```
+For any further styling now use `inputDecorationTheme` or `inputDecorationBuilder` properties 
+of `FastForm` (see example app).
+
+* renames `decorator` property on `FastForm` to `inputDecorationBuilder` and changes its typedef
+
+**old**
+```dart
+typedef FastInputDecorator = InputDecoration Function(ThemeData theme, FastFormFieldState field);
+```
+
+**new**
+```dart
+typedef FastInputDecorationBuilder = InputDecoration Function(FastFormFieldState field);
+```
+For retrieving `ThemeData` now use `Theme.of(context)` within the function body.
+
 ## [10.0.0] - 01/29/2023
 
 * upgrades to Flutter `3.7.0`